Various changes from ECSS branch

See merge request lely_industries/lely-core!62

+++ include/lely/co/sdev.h
@@ -2,7 +2,7 @@
  * This header file is part of the CANopen library; it contains the static
  * device description declarations.
  *
- * @copyright 2016-2019 Lely Industries N.V.
+ * @copyright 2016-2020 Lely Industries N.V.
  *
  * @author J. S. Seldenthuis <jseldenthuis@lely.com>
  *
@@ -97,11 +97,11 @@ struct co_ssub {
 	/// The sub-object value.
 	union co_val val;
 	/// The access type.
-	unsigned long access : 5;
+	uint_least32_t access : 5;
 	/// A flag indicating if it is possible to map this object into a PDO.
-	unsigned long pdo_mapping : 1;
+	uint_least32_t pdo_mapping : 1;
 	/// The object flags.
-	unsigned long flags : 26;
+	uint_least32_t flags : 26;
 };
 
 #ifdef __cplusplus
